Refactor: move struct to map test to a file that matches where it is implemented
