+++ validator/unknown_attribute_test.go
@@ -65,26 +65,6 @@ func Test_UnknownAttribute(t *testing.T) {
 	}
 }
 
-func Test_StructToMap(t *testing.T) {
-	type testStruct struct {
-		Foo string
-		Bar int
-	}
-
-	s := testStruct{}
-
-	m, err := StructToMap(s)
-	if err != nil {
-		t.Fatal("expected", nil, "got", err)
-	}
-	if _, ok := m["Foo"]; !ok {
-		t.Fatal("expected", true, "got", false)
-	}
-	if _, ok := m["Bar"]; !ok {
-		t.Fatal("expected", true, "got", false)
-	}
-}
-
 func testExpectedForUnknownFields(t *testing.T) map[string]interface{} {
 	return map[string]interface{}{
 		"foo":   "foo",
